fix(precision): fix arbitrary precision problem
Sometimes, when we compare coordinates, we get 2 semantically equivalent
coordinates (e.g. .30000000004 and 0.3), but the `coordEqual` would
return false anyway because the coords are not exactly equals

I've made the hypothesis that nobody will ever need a precision greater
than 10 decimals.

The test `Handles implicite moveTo` was changed because the last 2
segment were semantically equivalent.